requirements.txt
Cython==0.28.5
cytoolz==0.9.0.1
opencv-python==3.4.0.12
pathos==0.2.2
Pillow==5.4.1
scikit-image==0.14.0
scikit-learn==0.19.2
scipy==1.0.1
sklearn==0.0
torch==0.4.0
torchfile==0.1.0
torchvision==0.2.1
tqdm==4.23.0
